Update to Hyper 0.10 and Serde 0.9.
