+++ docker-compose.yml
@@ -1,20 +1,22 @@
-etcd:
-  image: quay.io/coreos/etcd:v2.3.0
-  command: -advertise-client-urls=http://0.0.0.0:2379 -listen-client-urls=http://0.0.0.0:2379
-etcdsecure:
-  image: quay.io/coreos/etcd:v2.3.0
-  command: -advertise-client-urls=https://0.0.0.0:2379 -listen-client-urls=https://0.0.0.0:2379 -trusted-ca-file=/ca.pem -cert-file=/cert.pem -key-file=/key.pem -client-cert-auth
-  volumes:
-    - ./tests/ssl/ca.pem:/ca.pem
-    - ./tests/ssl/server.pem:/cert.pem
-    - ./tests/ssl/server-key.pem:/key.pem
-rust:
-  image: jimmycuadra/rust
-  environment:
-    RUST_BACKTRACE: 1
-    RUST_TEST_THREADS: 1
-  links:
-    - etcd
-    - etcdsecure
-  volumes:
-    - .:/source
+version: "2"
+services:
+  etcd:
+    image: quay.io/coreos/etcd:v2.3.7
+    command: -advertise-client-urls=http://0.0.0.0:2379 -listen-client-urls=http://0.0.0.0:2379
+  etcdsecure:
+    image: quay.io/coreos/etcd:v2.3.7
+    command: -advertise-client-urls=https://0.0.0.0:2379 -listen-client-urls=https://0.0.0.0:2379 -trusted-ca-file=/ca.pem -cert-file=/cert.pem -key-file=/key.pem -client-cert-auth
+    volumes:
+      - ./tests/ssl/ca.pem:/ca.pem
+      - ./tests/ssl/server.pem:/cert.pem
+      - ./tests/ssl/server-key.pem:/key.pem
+  rust:
+    image: jimmycuadra/rust
+    environment:
+      RUST_BACKTRACE: 1
+      RUST_TEST_THREADS: 1
+    links:
+      - etcd
+      - etcdsecure
+    volumes:
+      - .:/source
